Make page title match actual attribute name
The page title was `remote_state`, but the actual attribute (not sure if that's the right word) is `terraform_remote_state`. This confused me and probably confuses others. So unless there's actually a `remote_state` that's different from `terraform_remote_state`, it makes sense to have the title actual follow the name you'd use in the config